move get_xla_options and get_megascale_options to pathways_utils.py
